#134 refactor: index export 구조 개선
