+++ src/Pages/InvitePage.tsx
@@ -4,8 +4,8 @@ import Button from '@/components/common/Button/Button';
 import Form from '@/components/common/Form';
 import IconWrapper from '@/components/common/IconWrapper';
 import { checkedStyle, checkIconStyle } from '@/components/common/Todo/TodoList';
-import useInviteMember from '@/hooks/mutations/useInviteMember';
-import useToastHandler from '@/hooks/useToastHandler';
+import { useToastHandler } from '@/hooks';
+import { useInviteMember } from '@/hooks/mutations';
 import { useSquadStore } from '@/stores';
 import { SearchMemberResponse } from '@/types';
 import { css, Theme, useTheme } from '@emotion/react';
