+++ src/Pages/SquadPage.tsx
@@ -1,6 +1,6 @@
 import Button from '@/components/common/Button/Button';
 import { useCreateSquad } from '@/hooks/mutations';
-import { squadQueryOptions } from '@/hooks/queries/useSquad';
+import { squadQueryOptions } from '@/hooks/queries';
 import { useToastStore } from '@/stores';
 import { breakpoints, mobileMediaQuery, pcMediaQuery } from '@/styles/breakpoints';
 import { Squad } from '@/types/squad';
