+++ src/components/common/Member/SelectMemberList.tsx
@@ -5,8 +5,8 @@ import MemberProfile from '@/components/common/Member/MemberProfile';
 import { ActionPrompt } from '@/components/common/Modal/ModalContents';
 import { ROLE } from '@/constants/role';
 import { MEMBER_STYLE_TYPE } from '@/constants/squad';
-import { squadDetailQueryOptions, squadKeys } from '@/hooks/queries/useSquad';
-import { useModal } from '@/hooks/useModal';
+import { useModal } from '@/hooks';
+import { squadDetailQueryOptions, squadKeys } from '@/hooks/queries';
 import { useToastStore } from '@/stores';
 import { SquadMember } from '@/types';
 import handleKeyDown from '@/utils/handleKeyDown';
