+++ src/components/common/Member/SidebarMemberList.tsx
@@ -4,9 +4,9 @@ import MemberProfile from '@/components/common/Member/MemberProfile';
 import { commonButtonStyle } from '@/components/common/Sidebar';
 import { ROLE } from '@/constants/role';
 import { MEMBER_STYLE_TYPE } from '@/constants/squad';
+import { useUserCookie } from '@/hooks';
 import { useRemoveUserFromSquad } from '@/hooks/mutations';
-import { squadKeys } from '@/hooks/queries/useSquad';
-import useUserCookie from '@/hooks/useUserCookie';
+import { squadKeys } from '@/hooks/queries';
 import { useSquadStore, useToastStore } from '@/stores';
 import { mobileMediaQuery, pcMediaQuery } from '@/styles/breakpoints';
 import { SquadMember } from '@/types';
